+++ src/languages/_french.json
@@ -20,7 +20,7 @@
       "tab_interactions": "Interactions",
       "tab_interactions_tooltip": "Interactions de smart-contracts",
       "tab_investments": "Investissements",
-      "tab_positions": "Positions",
+      "tab_positions": "Positions :)",
       "tab_savings": "Epargnes",
       "tab_transactions": "Transactions",
       "tab_transactions_tooltip": "Transactions et transferts de tokens",
